refactor: interface segregation
