+++ src/interfaces/post.interface.ts
@@ -1,0 +1,5 @@
+export default interface Post {
+  slug: string
+  title: string
+  html: string
+}
