+++ src/interfaces/tab.interface.ts
@@ -1,0 +1,6 @@
+export interface Tab {
+  href: string
+  title: string
+  segment?: string
+  prefetch: boolean
+}
